requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-celery-results==2.0.1
django-redis==4.12.1
django-user-agents==0.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
flower==0.9.7
humanize==3.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.8
pkg-resources==0.0.0
prometheus-client==0.8.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
tornado==6.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.3
user-agents==2.2.0
vine==1.3.0
